+++ src/styles/theme/useTheme.ts
@@ -1,7 +1,7 @@
 import { useContext } from "react";
 import { ThemeContext } from "./ThemeContext";
 import { Theme, UseThemeResult } from "./types";
-import { LOCAL_STORAGE_THEME_KEY } from "utils/constants";
+import { LOCAL_STORAGE_THEME_KEY } from "../../utils/constants";
 
 export function useTheme(): UseThemeResult {
   const { theme, setTheme } = useContext(ThemeContext);
